Split debugTextStyle/builders implementations
