+++ tests/unit/labelTests.cpp
@@ -1,6 +1,7 @@
 #include "catch.hpp"
 #include "tangram.h"
 #include "labels/label.h"
+#include "style/textStyle.h"
 #include "labels/textLabel.h"
 #include "glm/mat4x4.hpp"
 #include "glm/gtc/matrix_transform.hpp"
